// client/src/app/shared/models/Candidate.ts
export interface ICandidate {
  id: number;
  firstName: string;
  lastName: string;
  address1: string;
  address2: string;
  address3: string;
  address4: string;
  address5: string;
  contactNumber: string;
  email: string;
  accoutNumber: string;
  accoutName: string;
  sortCode: string;
  photoUrl: string;
  appUserId: string;
  grade: string;
  gradeId: number;
}
export class Candidate {
    id: number;
    firstName: string;
    lastName: string;
    address1: string;
    address2: string;
    address3: string;
    address4: string;
    address5: string;
    contactNumber: string;
    email: string;
    accoutNumber: string;
    accoutName: string;
    sortCode: string;
    photoUrl: string;
    appUserId: string;
    grade: string;
    gradeId: number;
  }